refactor: fedimint-gateway-client
