+++ fedimint-testing/src/fixtures.rs
@@ -16,8 +16,9 @@ use fedimint_core::envs::BitcoinRpcConfig;
 use fedimint_core::module::registry::ModuleRegistry;
 use fedimint_core::task::{MaybeSend, MaybeSync};
 use fedimint_core::util::SafeUrl;
+use fedimint_gateway_common::LightningMode;
 use fedimint_gateway_server::client::GatewayClientBuilder;
-use fedimint_gateway_server::config::{LightningMode, LightningModuleMode};
+use fedimint_gateway_server::config::LightningModuleMode;
 use fedimint_gateway_server::Gateway;
 use fedimint_lightning::{ILnRpcClient, LightningContext};
 use fedimint_logging::TracingSetup;
