+++ gateway/fedimint-gateway-client/src/config_commands.rs
@@ -1,8 +1,8 @@
 use clap::Subcommand;
 use fedimint_core::config::FederationId;
 use fedimint_core::Amount;
+use fedimint_gateway_client::GatewayRpcClient;
 use fedimint_gateway_common::{ConfigPayload, SetFeesPayload};
-use fedimint_gateway_server::rpc::rpc_client::GatewayRpcClient;
 
 use crate::print_response;
