+++ gateway/fedimint-gateway-client/src/lib.rs
