--- gateway/fedimint-gateway-server/src/rpc/mod.rs
@@ -1,2 +1,0 @@
-pub mod rpc_client;
-pub mod rpc_server;
