+++ gateway/fedimint-gateway-server/src/rpc_server.rs
