+++ misc/git-hooks/pre-commit
@@ -84,7 +84,7 @@ function check_check_forbidden_dependencies() {
       >&2 echo "fedimint-server/Cargo.toml must not depend on modules"
       return 1
     fi
-    if grep -E "fedimint-[a-zA-Z0-9]+-(server|client)" fedimint-testing/Cargo.toml | grep -v "fedimint-api-client" >&2 ; then
+    if grep -E "fedimint-[a-zA-Z0-9]+-(server|client)" fedimint-testing/Cargo.toml | grep -v -E "fedimint-api-client|fedimint-gateway-*" >&2 ; then
       >&2 echo "fedimint-testing/Cargo.toml must not depend on modules"
       return 1
     fi
